+++ packages/asc-ui/src/components/Accordion/index.ts
@@ -1,0 +1,3 @@
+export { default as AccordionStyle } from './AccordionStyle'
+export { default as Accordion } from './Accordion'
+export { default as AccordionWrapper } from './AccordionWrapper'